requirements-api.txt
# FastAPI Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Existing dependencies
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

